pyproject.toml
[project]
name = "cm-mcp"
version = "0.1.0"
description = "MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp",
    "asyncpg",
    "pyyaml",
    "python-dotenv",
    "mcp>=1.26.0",
    "openai>=2.16.0",
    "aiohttp>=3.13.3",
    "uvicorn>=0.40.0",
    "py-key-value-aio[redis]>=0.3.0",
    "langchain>=1.2.14",
    "langchain-openai>=1.1.12",
    "langchain-google-genai>=4.2.1",
    "supervisor>=4.3.0",
]
